+++ src/main/java/com/moltenwolfcub/crafted_cuisine/block/AutoBlowTorchBlock.java
@@ -2,7 +2,6 @@ package com.moltenwolfcub.crafted_cuisine.block;
 
 import java.util.stream.Stream;
 
-import net.minecraft.world.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -11,7 +10,14 @@ import com.moltenwolfcub.crafted_cuisine.init.AllBlockEntities;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.Container;
+import net.minecraft.world.Containers;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -119,6 +125,13 @@ public class AutoBlowTorchBlock extends BaseEntityBlock {
         }
 
         return InteractionResult.SUCCESS;
+    }
+
+    @Override
+    public void setPlacedBy(Level level, BlockPos blockPos, BlockState blockState, LivingEntity livingEntity, ItemStack itemStack) {
+        if (itemStack.hasCustomHoverName() && level.getBlockEntity(blockPos) instanceof AutoBlowTorchBlockEntity blockEntity) {
+            blockEntity.setCustomName(itemStack.getHoverName());
+        }
     }
 
     @Nullable