+++ src/main/java/com/moltenwolfcub/crafted_cuisine/block/CarameliserBlock.java
@@ -1,6 +1,5 @@
 package com.moltenwolfcub.crafted_cuisine.block;
 
-import net.minecraft.world.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -10,7 +9,14 @@ import com.moltenwolfcub.crafted_cuisine.init.AllBlockEntities;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.world.Container;
+import net.minecraft.world.Containers;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -108,6 +114,13 @@ public class CarameliserBlock extends BaseEntityBlock {
         }
 
         return InteractionResult.SUCCESS;
+    }
+
+    @Override
+    public void setPlacedBy(Level level, BlockPos blockPos, BlockState blockState, LivingEntity livingEntity, ItemStack itemStack) {
+        if (itemStack.hasCustomHoverName() && level.getBlockEntity(blockPos) instanceof CarameliserBlockEntity blockEntity) {
+            blockEntity.setCustomName(itemStack.getHoverName());
+        }
     }
 
     @Nullable